langchain_health_beahviour_webapp/requirements.txt
openai
langchain
langchain-experimental
python-dotenv
pandas
scikit-learn
seaborn
matplotlib
streamlit
streamlit-toggle-switch
tabulate
tiktoken
duckduckgo-search
xmltodict